% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loss.R
\name{get.loss}
\alias{get.loss}
\title{A function that calculates the loss/cost}
\usage{
get.loss(y, pred, type)
}
\arguments{
\item{y}{The response vector.}

\item{pred}{The predicted values for the response.}

\item{type}{A string indicating the type of regression model (linear or binomial).}
}
\value{
The loss of the input vectors.
}
\description{
A function that calculates the loss/cost
}
