% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_group.R
\name{process.group}
\alias{process.group}
\title{Process groupings for a sparse group penalty}
\usage{
process.group(group, group.weight)
}
\arguments{
\item{group}{A vector that specifies the group membership of each variable in X.}

\item{group.weight}{A vector specifying weights that are multiplied by the group penalty to account for different group sizes.}
}
\value{
A structure containing the prepared group structure and, as an attribute, its labels and group weights.
}
\description{
A function that checks the group information for possible errors and processes it.
}
