% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_X.R
\name{process.X}
\alias{process.X}
\title{Process X for a sparse group penalty}
\usage{
process.X(X, group)
}
\arguments{
\item{X}{The design matrix without intercept with the variables to be selected.}

\item{group}{A vector that specifies the group membership of each variable in X.}
}
\value{
A list containing:
\describe{
\item{X}{The standardized design matrix X.}
\item{vars}{The variable names of the matrix.}
\item{center}{The center of the variables before the transformation.}
\item{scale}{The scale of the variables before the transformation.}
}
}
\description{
A function that checks the design matrix X for possible errors and scales it.
}
