% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_y.R
\name{process.y}
\alias{process.y}
\title{Process y for a sparse group penalty}
\usage{
process.y(y, type)
}
\arguments{
\item{y}{The response vector.}

\item{type}{A string indicating the type of regression model (linear or binomial).}
}
\value{
The verified response vector y.
}
\description{
A function that checks the response vector y for possible errors.
}
