% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_Z.R
\name{process.Z}
\alias{process.Z}
\title{Process Z for a sparse group penalty}
\usage{
process.Z(Z)
}
\arguments{
\item{Z}{The design matrix of the variables to be included in the model without penalization.}
}
\value{
A list containing:
\describe{
\item{Z}{The standardized design matrix Z.}
\item{vars}{The variable names of the matrix.}
\item{center}{The center of the variables before the transformation.}
\item{scale}{The scale of the variables before the transformation.}
}
}
\description{
A function that checks the design matrix Z for possible errors and scales it.
}
