% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGPR-package.R
\docType{package}
\name{SGPR-package}
\alias{SGPR}
\alias{SGPR-package}
\title{SGPR: Sparse Group Penalized Regression for Bi-Level Variable Selection}
\description{
Fits the regularization path of regression models (linear and logistic) with additively combined penalty terms. All possible combinations with Least Absolute Shrinkage and Selection Operator (LASSO), Smoothly Clipped Absolute Deviation (SCAD), Minimax Concave Penalty (MCP) and Exponential Penalty (EP) are supported. This includes Sparse Group LASSO (SGL), Sparse Group SCAD (SGS), Sparse Group MCP (SGM) and Sparse Group EP (SGE). For more information, see Buch, G., Schulz, A., Schmidtmann, I., Strauch, K., & Wild, P. S. (2024) \doi{10.1002/bimj.202200334}.
}
\author{
\strong{Maintainer}: Gregor Buch \email{buchgregor@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Andreas Schulz [thesis advisor]
  \item Irene Schmidtmann [thesis advisor]
  \item Konstantin Strauch [thesis advisor]
  \item Philipp Wild [thesis advisor]
}

}
\keyword{internal}
